+++ projects/react/components/Header/Header.js
@@ -10,10 +10,10 @@ import { toggleModalFullscreen } from '../../utils/full-screen-handle';
 export default class extends Component {
   render() {
     const {
-      activeTab, onRevert, apply, onClose, showGoBackBtn, processWithCloudService, processWithFilerobot,
+      activeTab, onRevert, apply, onClose, processWithCloudService, processWithFilerobot,
       handleSave, activeBody, t, config
     } = this.props;
-    const { tools } = config;
+    const { tools, showGoBackBtn } = config;
     const isOneTool = tools.length === 1;
     const filteredName = activeTab === 'rotate' ? 'orientation' : activeTab;
     const onFinishButtonLabel = (!processWithCloudService && !processWithFilerobot) ? t['toolbar.download'] : t['toolbar.save'];
@@ -29,12 +29,12 @@ export default class extends Component {
 
         <ToolbarWrapper overlayYHidden={activeTab !== 'watermark'}>
           <ActionsWrapper>
-            <LeftActions hide={!activeTab}>
-              <CancelBtn onClick={isOneTool ? onClose : onRevert} sm default fullSize>
+            <LeftActions>
+              <CancelBtn hide={!activeTab} onClick={isOneTool ? onClose : onRevert} sm default fullSize>
                 {t[`toolbar.cancel`]}
               </CancelBtn>
               {showGoBackBtn &&
-              <CancelBtn hide={activeTab} onClick={onClose} sm default fullSize>{t[`toolbar.go_back`]}</CancelBtn>}
+              <CancelBtn onClick={onClose} sm default fullSize>{t[`toolbar.go_back`]}</CancelBtn>}
             </LeftActions>
 
             {activeBody === 'preview' &&